+++ src/svg.js
@@ -135,7 +135,13 @@ SVG.prepare = function() {
   // Create parser object
   SVG.parser = {
     body: body || document.documentElement
-  , draw: draw.style('opacity:0;position:absolute;left:-100%;top:-100%;overflow:hidden')
+  , draw: draw.style({
+      opacity:0,
+      position:'absolute',
+      left:'-100%',
+      top:'-100%',
+      overflow:'hidden'
+    })
   , poly: draw.polyline().node
   , path: draw.path().node
   , native: SVG.create('svg')